finding dirac-test-job in extension

+++ tests/Jenkins/utilities.sh
@@ -371,8 +371,8 @@ installDIRAC() {
 ##############################################################################
 # This function submits a job or more (it assumes a DIRAC client is installed)
 # it needs the following environment variables:
-# ${DIRACUSERDN} for the DN of the user used to submit the job
-# ${DIRACUSERROLE} for the role of the proxy of the user used to submit the job
+# $DIRACUSERDN for the DN of the user used to submit the job
+# $DIRACUSERROLE for the role of the proxy of the user used to submit the job
 # $DIRACSETUP for the setup
 
 submitJob() {
@@ -384,11 +384,26 @@ submitJob() {
   fi
 
   export PYTHONPATH=${TESTCODE}:${PYTHONPATH}
-  #Get a proxy and submit the job: this job will go to the certification setup, so we suppose the JobManager there is accepting jobs
+  # Get a proxy and submit the job: this job will go to the certification setup, so we suppose the JobManager there is accepting jobs
+
+  # check if errexit mode is set and disabling as the component may not exist
+  save=$-
+  if [[ $save =~ e ]]; then
+    set +e
+  fi
   getUserProxy #this won't really download the proxy, so that's why the next command is needed
+  # re-enabling it
+  if [[ ${save} =~ e ]]; then
+    set -e
+  fi
+
   cp "${TESTCODE}/DIRAC/tests/Jenkins/dirac-proxy-download.py" "."
   python dirac-proxy-download.py "${DIRACUSERDN}" -R "${DIRACUSERROLE}" -o /DIRAC/Security/UseServerCertificate=True -o /DIRAC/Security/CertFile=/home/dirac/certs/hostcert.pem -o /DIRAC/Security/KeyFile=/home/dirac/certs/hostkey.pem -o /DIRAC/Setup="${DIRACSETUP}" -ddd
-  cp "${TESTCODE}/DIRAC/tests/Jenkins/dirac-test-job.py" "."
+  if [[ -f "${TESTCODE}/${VO}DIRAC/tests/Jenkins/dirac-test-job.py" ]]; then
+    cp "${TESTCODE}/${VO}DIRAC/tests/Jenkins/dirac-test-job.py" "."
+  else
+    cp "${TESTCODE}/DIRAC/tests/Jenkins/dirac-test-job.py" "."
+  fi
   python dirac-test-job.py -o "/DIRAC/Setup=${DIRACSETUP}" "${DEBUG}"
 
   echo '==> Done submitJob'
